Refactor IngredientRepository to handle nutrition CRUD Remove unused nutrition repository test file and class
